JUnit agregado con 2 test para cada método

+++ src/Calculator.java
@@ -66,7 +66,7 @@ public class Calculator {
     public static int multiplicar(int a, int b){
         return a*b;
     }
-    public static float dividir(int a, int b){
+    public static float dividir(float a, float b){
         return a/b;
     }
